Splitted views into mutiple files, got rid of forms.py
